requirements.txt
certifi==2019.11.28
chardet==3.0.4
docutils==0.16
fuzzywuzzy==0.17.0
idna==2.8
Kivy==2.0.0rc1
kivy-deps.glew==0.2.0
kivy-deps.glew-dev==0.2.0
kivy-deps.sdl2==0.2.0
kivy-deps.sdl2-dev==0.2.0
Kivy-Garden==0.1.4
kivymd==0.103.0
Pillow==7.0.0
Pygments==2.5.2
python-Levenshtein==0.12.0
PyYAML==5.3
requests==2.22.0
urllib3==1.25.8
